rename model_server to base
