+++ chitra/serve/__init__.py
@@ -1,3 +1,3 @@
 from chitra.serve.api import API, create_api
 from chitra.serve.app import GradioApp
-from chitra.serve.model_server import ModelServer
+from chitra.serve.base import ModelServer
